231229 이성주: GNB 기능 수정완료
